+++ app.js
@@ -2,29 +2,15 @@ const express = require("express");
 const app = express();
 const path = require("path");
 const newsRoutes = require("./routes/index");
+const eachNewsRoutes = require("./routes/news");
 
 app.set("view engine", "ejs");
 app.set("views", path.join(__dirname, "./views"));
 
 app.use(express.static(path.join(__dirname, "public")));
 
-app.get("/", newsRoutes);
-
-const fs = require("fs");
-
-app.get("/news/:id", async (req, res) => {
-  try {
-    const { id } = req.params;
-    const filePath = path.join(__dirname, "data/headlines.json");
-    const data = fs.readFileSync(filePath, "utf-8");
-    const newsData = JSON.parse(data);
-    const newsArticles = newsData.articles;
-    res.render("selectedNews", { newsArticles, id });
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    res.status(500).json({ error: "Failed to fetch data" });
-  }
-});
+app.use("/", newsRoutes);
+app.use("/news", eachNewsRoutes);
 
 app.listen(8080, () => {
   console.log("listening on port 8080!!");
